+++ packages/flutter_formio/example/test/widget_integration_test.dart
@@ -1,6 +1,6 @@
 // ignore_for_file: avoid_print
 
-import 'package:flutter_formio/flutter_formio.dart';
+import 'package:formio/formio.dart';
 
 /// Comprehensive Widget Integration Test
 ///