+++ packages/flutter_formio/lib/src/models/file_typedefs.dart
@@ -2,7 +2,7 @@
 library;
 
 import 'package:flutter/material.dart';
-import 'package:formio/flutter_formio.dart';
+import 'package:formio/formio.dart';
 import 'package:formio_api/formio_api.dart' show FileData;