+++ packages/flutter_formio/lib/src/widgets/components/columns_component.dart
@@ -6,7 +6,7 @@
 library;
 
 import 'package:flutter/material.dart';
-import 'package:formio/flutter_formio.dart';
+import 'package:formio/formio.dart';
 
 import '../form_data_provider.dart';