+++ packages/flutter_formio/lib/src/widgets/components/datetime_component.dart
@@ -4,7 +4,7 @@
 library;
 
 import 'package:flutter/material.dart';
-import 'package:formio/flutter_formio.dart';
+import 'package:formio/formio.dart';
 import 'package:intl/intl.dart';
 
 class DateTimeComponent extends StatefulWidget {