+++ packages/flutter_formio/lib/src/widgets/components/dynamic_wizard_component.dart
@@ -8,7 +8,7 @@
 library;
 
 import 'package:flutter/material.dart';
-import 'package:formio/flutter_formio.dart';
+import 'package:formio/formio.dart';
 
 
 class DynamicWizardComponent extends StatefulWidget {