+++ packages/flutter_formio/lib/src/widgets/form_renderer.dart
@@ -10,7 +10,7 @@ library;
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:formio/flutter_formio.dart';
+import 'package:formio/formio.dart';
 
 import 'form_data_provider.dart';