+++ packages/flutter_formio/test/core/cross_field_validation_test.dart
@@ -1,6 +1,6 @@
 // ignore_for_file: depend_on_referenced_packages
 
-import 'package:formio/flutter_formio.dart';
+import 'package:formio/formio.dart';
 import 'package:flutter_test/flutter_test.dart';