+++ packages/formio_api/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [2.0.2] - 2026-01-13
+
+### Added
+
+- Added DataSourceService to main library file
+
 ## [2.0.1] - 2026-01-13
 
 ### Fixed